requirements.txt
PyYAML >= 5.3
UV